chore: refactored controllers into a single folder
