+++ src/controllers/orders/dto/order.dto.ts
@@ -1,6 +1,6 @@
 import { Expose, Transform } from 'class-transformer';
-import { OrderedIngredients } from 'src/ordered-ingredients/ordered-ingredients.entity';
-import { User } from 'src/users/users.entity';
+import { OrderedIngredients } from 'src/controllers/ordered-ingredients/ordered-ingredients.entity';
+import { User } from 'src/controllers/users/users.entity';
 
 export class OrderDTO {
   @Expose()
