+++ src/controllers/orders/orders.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { OrderedIngredientsModule } from 'src/ordered-ingredients/ordered-ingredients.module';
-import { UsersModule } from 'src/users/users.module';
+import { OrderedIngredientsModule } from 'src/controllers/ordered-ingredients/ordered-ingredients.module';
+import { UsersModule } from 'src/controllers/users/users.module';
 import { OrdersController } from './orders.controller';
 import { Order } from './orders.entity';
-import { OrdersService } from '../services/orders.service';
+import { OrdersService } from 'src/services/orders/orders.service';
 
 @Module({
   imports: [
