+++ src/controllers/users/auth.controller.ts
@@ -8,12 +8,11 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { CreateUserDTO, SignInDTO, UserDTO } from './dtos/dto';
-import { AuthService } from '../services/auth/auth.service';
-import { CurrentUser } from '../core/decorators/current-user.decorators';
+import { AuthService } from 'src/services/auth/auth.service';
+import { CurrentUser } from 'src/core/decorators/current-user.decorators';
 import { LocalAuthGuard } from 'src/core/guards/local-auth.guard';
 import { JwtAuthGuard } from 'src/core/guards/jwt-auth.guard';
 import { AuthGuard } from 'src/core/guards/auth.guard';
-import { serialize } from 'class-transformer';
 import { Serialize } from 'src/core/interceptors/serialize.inteceptor';
 
 @Controller('auth')
