+++ src/controllers/users/users.module.ts
@@ -2,7 +2,7 @@ import { MiddlewareConsumer, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersController } from './users.controller';
 import { User } from './users.entity';
-import { UsersService } from '../services/users.service';
+import { UsersService } from 'src/services/users/users.service';
 import { CurrentUserMiddleWare } from 'src/core/middleware/current-user.middleware';
 import { AuthService } from 'src/services/auth/auth.service';
