+++ src/core/interceptors/current-user.interceptors.ts
@@ -4,7 +4,7 @@ import {
   CallHandler,
   Injectable,
 } from '@nestjs/common';
-import { UsersService } from '../../services/users.service';
+import { UsersService } from '../../services/users/users.service';
 
 @Injectable()
 export class CurrentUserInterceptor implements NestInterceptor {
