+++ src/services/ingredients/ingredients.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Ingredient } from '../ingredients/ingredients.entity';
+import { Ingredient } from '../../controllers/ingredients/ingredients.entity';
 
 @Injectable()
 export class IngredientsService {
